# docs/api.yaml
openapi: 3.0.3
info:
  title: Numeric bank api specs
  description: This document specify the endpoints we'll use for our numeric bank web application
  version: 3.6.0
servers:
  - url: "http://localhost:8080"
...
paths:
...
  /transactions:
    get:
      tags:
        - Transaction
      summary: get list of Transaction
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseBody"
                properties:
                  body:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

    post:
      tags:
        - Transaction
      summary: create new transaction
      requestBody:
        $ref: "#/components/requestBodies/TransactionBody"
      responses:
        200:
          description: the created transaction
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseBody"
                properties:
                  body:
                    $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"     
    put:
      tags:
        - Transaction
      summary: update an existent transaction
      requestBody:
        $ref: "#/components/requestBodies/UpdateTransactionBody"
      responses:
        200:
          description: the updated transaction
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseBody"
                properties:
                  body:
                    $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500" 
  /accounts/{idAccount}/transactions:
    get:
      tags:
        - Transaction
      summary: get list of Transaction by the id_account
      description: You can filter it with id of the transaction, transaction_datetime and the transaction_type
      parameters:
        - $ref: '#/components/parameters/IdAccount'
        - $ref: '#/components/parameters/IdTransaction'
        - $ref: '#/components/parameters/TransactionDatetime'
        - $ref: '#/components/parameters/TransactionType'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseBody"
                properties:
                  body:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"              
  /transactions/{id}:
    get:
      tags:
        - Transaction
      summary: get Transaction by his id
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        200:
          description: the corresponding Transaction
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseBody"
                properties:
                  body:
                    $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500" 
    delete:
      tags:
        - Transaction
      summary: delete a transaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        200:
          description: the deleted Transaction
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseBody"
                properties:
                  body:
                    $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500" 
...
components:
...
  requestBodies:
...
    TransactionBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/crudTransaction" 
    UpdateTransactionBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Transaction"        
...
  parameters:
    IdAccount:
      name: idAccount
      in: path
      required: true
      schema:
        type: integer
...
    IdTransaction:
      name: idTransaction
      in: query
      required: false
      schema:
        type: integer    
    TransactionDatetime:
      name: TransactionDatetime
      in: query
      required: false
      schema:
        type: string
        format: datetime       
    TransactionType:
      name: transactionType
      in: query
      required: false
      schema:   
        $ref: '#/components/schemas/transactionType' 
...
    TransactionId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    400:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BadRequestException"
    403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ForbiddenException"
    404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResourceNotFoundException"
    429:
      description: Too many Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TooManyRequestsException"
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InternalServerException"

  ### schemas components
  schemas:
    CommonId:
      type: object
      properties:
        id:
          type: integer
          format: int64
...
    transactionType:
      type: string
      enum:
        - CREDIT
        - DEBIT 
    crudTransaction:
      type: object
      properties:
        id_account:
          type: integer
          format: int64
        transaction_datetime:
          type: string 
          format: datetime
        amount:
          type: integer
          format: double
        type:
          $ref: '#/components/schemas/transactionType'
    Transaction:
      allOf:
        - $ref: '#/components/schemas/CommonId'
        - $ref: '#/components/schemas/crudTransaction'
...
    ResponseStatus:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      example:
        code: 200
        message: OK
    ResponseBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ResponseStatus"
        body:
          type: object

    ### EXCEPTIONS
    BadRequestException:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
      example:
        message: BadRequest
        code: 400
        body: {}
...
    ForbiddenException:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
      example:
        code: 403
        message: Forbidden
        body: {}
    ResourceNotFoundException:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
      example:
        code: 404
        message: ResourceNotFound
        body: {}
    TooManyRequestsException:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
      example:
        code: 429
        message: TooManyRequests
        body: {}
    InternalServerException:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
      example:
        code: 500
        type: InternalServerError
        body: {}